Define eddy4R.york Parameters

Create para list, use arguments to overide defaults

def.para(
  DirInp,
  DirWrk,
  runID,
  analysis,
  fileMask,
  fileDuration,
  DirOut = NULL,
  writeCompressedData = TRUE,
  siteName = NULL,
  subDir = c("none", "monthly", "daily")[1],
  species = NULL,
  speciesRatioName = NULL,
  speciesFluxName = NULL,
  unitList = NULL,
  freq = 5,
  fileNames = NULL,
  filePaths = NULL,
  filePattern = "*.csv",
  tz = "UTC",
  dateFormat = "%Y-%m-%d %H:%M:%OS",
  varsRequired = c("unixTime", "distZaxsMeas", "presAtm", "distZaxsAbl"),
  varsCritical = c("veloXaxs", "veloYaxs", "veloZaxs", "tempAir"),
  AlgBase = "trnd",
  idepVar = "unixTime",
  aggregationDuration = 3600,
  missingThreshold = 0.1,
  missingMethod = c("drop", "mean")[1],
  lagNOc = F,
  anemometerType = c("other", "CSAT3"),
  anemometerOffset = NULL,
  veloXaxs = "+veloXaxs",
  veloYaxs = "+veloYaxs",
  veloZaxs = "+veloZaxs",
  wBoost = F,
  lagApplyCorrection = T,
  lagNgtvPstv = c("n", "p", "np")[1],
  lagDefaults = -10,
  lagApplyRangeLimit = F,
  lagRangeLimit = NULL,
  lagVars = NULL,
  lagMax = 40 * freq,
  freqThsh = (1/(2 * lagMax/freq)),
  despike = T,
  despikeThreshold = NULL,
  despikeVars = NULL,
  MethRot = c("single", "double", "planarFit")[2],
  plnrFitCoef = NULL,
  plnrFitType = c("simple", "time", "wind")[1],
  lat,
  PltfEc = "towr",
  ZoneUtm = data.frame(Zone = 30, Estg = 698478, Nthg = 5711690),
  fileFirstStart = NULL,
  fileLastStart = NULL,
  spectralTaperingWeight = 0.5,
  footprintResolutionM = 100,
  footprintCumulativeThreshold = 0.8,
  wavelet_ec = FALSE,
  wavelet_av = "mean",
  wavelet_win = 5,
  wavelet_agg = 1,
  wavelet_cores = 2,
  wavelet_maxscale = 60,
  ...
)

Arguments

DirInp

must be supplied - relative to DirWrk. If input files are in subfolders (e.g. DirWrk/in/yyyy/mm) , specify just the top level here ('in')

DirWrk

root of the data directory

runID

Used in file nameing

analysis

Used in file nameing

fileMask

string mask that can be passed to as.POSIXct(format = mask) to decode the date in the fileName

fileDuration

expected duration of a complete input file (Hz)

DirOut

by default is created as file.path(DirWrk,"out",siteName, runID, analysis), can be overridden here.

writeCompressedData

TRUE/FALSE should the larger outputs that need compression ("ACF","foot", "spec", "base", "conv", "data","diff") outputs be written to disk. They will be compressed using gzip

siteName

Used in file nameing

subDir

one of c("none", "monthly", "daily") - default none. Should the outputs be split into monthly or daily subdirectories

species

vector of species names. Match with IntlNatu naming

speciesRatioName

created automatically from species. name of gas species following the rtioMoleDry format

speciesFluxName

created automatically from species. name of gas species following the flux format

unitList

defaults to default_unit_list() plus species field based off of species argument speices is a list with names that match speciesRatioName describing the units per species. Defaults to mol mol-1Dry

freq

data aquisition frequency of input data

fileNames

vector of file names, created as basename(filePaths). fileMask should operate on this.

filePaths

vector of file paths, from which fileNames can be created using basename(filePaths). By default this is populated by running the find system command on the DirInp, in the format find DirInp -type f -name *filePattern

filePattern

pattern to help filter input directory - default '*.csv'.

tz

timezone, default "UTC".

dateFormat

default %Y-%m-%d %H:%M:%OS"

varsRequired

character vector of columns that must be nominally present to pass def.valid.input()

varsCritical

these must have greater than the missingThreshold to pass def.valid.input()

AlgBase

detrending method for def.base.ec "mean","trend","ord03"

idepVar

required if Algbase is trend or ord03. Name of column containing independant variable for determining the base function. Column gets duplicated to a column called "idep" use use in eddy4R.turb::def.stat.sta.diff. Default unixTime.

aggregationDuration

flux aggregation period (s)

missingThreshold

decimal percentage of missing data threshold per file

missingMethod

how should missing data be handeled if it is less than the threshold. "drop","mean"

lagNOc

when processing NOx, is NO2 actually NOc and therfore CE applied after lagging. T/F

anemometerType

"other" or "CSAT3". In the case of CSAT 3, use the def.met.body() function. Otherwise use def.vect.orie().
For CSAT 3, anemometer offset should be positve from north. For other, anemomet offset should be +ve for clockwise correction,
negative for anticlockwise

anemometerOffset

rotation angle for uv plane of wind vectors (degrees)

veloXaxs

wind vector definition for wrap.anem.cor

veloYaxs

wind vector definition for wrap.anem.cor

veloZaxs

wind vector definition for wrap.anem.cor

wBoost

should w boost correction for certain Gill Anemometers be applied? Default false

lagApplyCorrection

Should lag correction be undertaken? T/F

lagNgtvPstv

what "direction" can lag occur? "n","p","np"

lagDefaults

numeric value or vector to apply to lag if ccf falls outside of range or is not used. If this is NA, then the lag will be the value of maxmimum absolute ACF within the given lag ranges

lagApplyRangeLimit

Should a window be applied to the ccf lag?

lagRangeLimit

vector or list of vectors containing boundary for acceptable lags

lagVars

created from species plus temperature and water vapour

lagMax

maximum lag range in timesteps (Hz) to search for a lag over. Defaults to 40*freq

freqThsh

cutoff frequency for the low end of the high pass filter. defaults to 1 / (2 * lagMax / freq)

despike

Should despiking be undertaken? T/F

despikeThreshold

numeric vector of standard deviations for temperature and water vapour, plus all species

despikeVars

created from species plus temperature and water vapour

MethRot

type of coordinate rotation of wind vectors to be performed. "single","double","planarFit"

plnrFitCoef

coefficients for planar fit calculation.
either numeric vector for plnrFitType = "simple" or data.frame for plnrFitType = "wind" or "time"

plnrFitType

type of planar fit, "simple", "date" or "wind". (character vector)

simple - numeric vector constant of coefficeients, or coefficeients that are controlled from the workflow. c(al,be,b0)

time - data.frame with columns date, al, be, b0. values with date nearest to mn$date are used

time - data.frame with columns PSI_uv, al, be, b0. values with date nearest to mn$PSI_uv are used

lat

latitude of tower location

PltfEc

switch for eddy4r functions between tower and aircraft. towr,airc

ZoneUtm

data.frame containing UTM info e.g data.frame(Zone=30, Estg=698478, Nthg=5711690)

fileFirstStart

used to aid def.avg - should be created internally

fileLastStart

used to aid def.avg - should be created internally

spectralTaperingWeight

tapering value to be supplied to eddy4R.turb::def.spec.fft.fwd() and ultimately stats::spec.taper(). value < 0.5 or FALSE Default 0.05 as per eddy4R.turb::def.spec.fft.fwd().

footprintResolutionM

Resolution of footprint model in meters (default 100)

footprintCumulativeThreshold

threshold for cumulative footprint extent (default 0.8)

wavelet_ec

supply argument true or false for conducting wavelet eddy-covariance

wavelet_av

supply argument "mean" or "band" for conducting wavelet eddy-covariance averaging.

wavelet_win

supply value for data window to use for wavelet averaging. Tower setup give value in minutes, e.g. 5. Aircraft give distance in meters.

wavelet_agg

supply value for data jump to use for wavelet averaging. Tower setup give value in minutes, e.g. 1. Aircraft gives distance in meters.

wavelet_cores

specify number of CPU cores to use for parallel wavelet averaging.

wavelet_maxscale

maximum wavlet scale to use for averaging. Default is 60 minutes. Give value in in minutes, e.g. 60.

...

supply arguments that have not been given an explicit definition yet here

Author

W. S. Drysdale