Wrap K04 Footprint model

Wraps the K04 (eddy4R.turb::def.foot.k04) footprint model to pass params and tidy output into data.frame. Outputs data.frame with x, y and value columns x and y are in m (based off of footprintResolutionM).

wrap.foot.k04(
  angZaxsErth,
  footprintResolutionM,
  veloYaxsHorSd,
  veloZaxsHorSd,
  veloFric,
  distZaxsMeasDisp,
  distZaxsAbl,
  distZaxsRgh,
  footprintCumulativeThreshold,
  univFunc
)

Arguments

angZaxsErth

wind direction to rotate the inertial footprint matrix

footprintResolutionM

Resolution of footprint model in meters (default 100)

veloYaxsHorSd

crosswind fluctuations

veloZaxsHorSd

vertical wind fluctuations

veloFric

friction velocity

distZaxsMeasDisp

height of measurement - displacement

distZaxsAbl

boundary layer height

distZaxsRgh

Roughness length

footprintCumulativeThreshold

threshold for cumulative footprint extent (default 0.8)

univFunc

integral over the stability-dependent universal function to make the log-wind-profile applicable to different atmospheric stratifications; from def.func.univ()

Author

W. S. Drysdale