UoY Lag definition wrapper

Wrapper for def.lag to use in UoY workflows. Returns data lag times and acf information

Usage

wrap.lag(
  eddy.data,
  lagVars,
  lagApplyRangeLimit,
  lagRangeLimit,
  lagDefaults,
  lagNOc,
  freq,
  freqThsh = NULL,
  speciesRatioName,
  lagMax,
  lagNgtvPstv
)

Arguments

eddy.data

eddy.data object

lagVars

created from species plus temperature and water vapour

lagApplyRangeLimit

Should a window be applied to the ccf lag?

lagRangeLimit

vector or list of vectors containing boundary for acceptable lags

lagDefaults

numeric value or vector to apply to lag if ccf falls outside of range or is not used. If this is NA, then the lag will be the value of maxmimum absolute ACF within the given lag ranges

lagNOc

when processing NOx, is NO2 actually NOc and therfore CE applied after lagging. T/F

freq

data aquisition frequency of input data

freqThsh

cutoff frequency for the low end of the high pass filter. defaults to 1 / (2 * lagMax / freq)

speciesRatioName

created automatically from species. name of gas species following the rtioMoleDry format

lagMax

maximum lag range in timesteps (Hz) to search for a lag over. Defaults to 40*freq

lagNgtvPstv

what "direction" can lag occur? "n","p","np"