Wrap Spectral Analysis

Computes the power and co-spectra using eddy4R.turb::def.spec.fft.fwd() and eddy4R.turb::def.spec.cosp(). Output is then tidied into data.frame containing measured, relative (0-1) and normalised (f*(U/(z-d))) frequency, as well as the power and cospectra, which are marked by the "spectrumType" column

wrap.spec(eddy.data, speciesRatioName, freq, spectralTaperingWeight = 0.05)

Arguments

eddy.data

time domain data including "veloXaxsInp", "veloYaxsInp", "veloZaxs" and "tempAir". It is expected that lag corrections and other QA/QC has been performed on these data already.

speciesRatioName

chemical species for flux calculations, can be NULL if no species are present.

freq

data aquisition frequency of input data

spectralTaperingWeight

tapering value to be supplied to eddy4R.turb::def.spec.fft.fwd() and ultimately stats::spec.taper(). value < 0.5 or FALSE Default 0.05 as per eddy4R.turb::def.spec.fft.fwd().

Author

W. S. Drysdale