Read eddy4R input

reads the files relevant for a given flux aggregation period TODO this can be substantially tidied

read_input(
  DirInp,
  dateFormat,
  filePaths,
  periodStartDate,
  periodEndDate,
  tz,
  freq,
  idepVar,
  PltfEc
)

Arguments

DirInp

input directory

dateFormat

format of date when reading csv for as.POSIXct() - reading from UNIX time to this format - default %Y-%m-%d %H:%M:%OS"

filePaths

files to aggregate

periodStartDate

start of aggregation period

periodEndDate

end of aggregation period

tz

time zone

freq

data acquisition frequency

idepVar

variable that could be used if Algbase in "trend" or "ord03" column gets duplicated to a column called "idep" used in eddy4R.turb::def.stat.sta.diff

PltfEc

platform either "towr" or "airc"

Author

W. S. Drysdale